% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_path.R
\name{extract_path}
\alias{extract_path}
\title{Extract Parts of a Path}
\usage{
extract_path(x, n = 1, fun = head, inds = NULL)
}
\arguments{
\item{x}{A vector of path(s).}

\item{n}{Corresponds to n in \code{\link[utils]{head}} & \code{\link[utils]{tail}}.
Gives 1 to n of the beginning (\code{fun = head}) or end (\code{fun = tail}) of string.
Note this is over rode if \code{inds} is supplied.}

\item{fun}{Either the \code{\link[utils]{head}} or \code{\link[utils]{tail}} function.}

\item{inds}{An index of elements of the path to use (for when you don't want to include
beginning or end of the path).}
}
\value{
Returns a partial path(s).
}
\description{
Extract portions of a path.
}
\examples{
extract_path(x)
extract_path(x, 2)
extract_path(x, -2)
extract_path(x, fun=tail)
extract_path(x, 2, fun=tail)
extract_path(x, inds=2:3)
x <- c(
    "C:/Users/trinker/Desktop/commencement_project",
    "C:\\\\Users\\\\trinker\\\\Desktop\\\\commencement_project"
)
}
\keyword{path}

