% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_lookup.R
\name{get_subfolder_path}
\alias{get_subfolder_path}
\title{Extract Paths Within a Folder}
\usage{
get_subfolder_path(listicle, subfolder)
}
\arguments{
\item{listicle}{A list of paths from\code{\link[base]{list.dirs}}.}

\item{subfolder}{A subfolder name to search in.}
}
\value{
Returns matching files (ignoring case) in subfolder.
}
\description{
Extract All Paths Within a Folder
}

