% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_match.R
\name{grep_match}
\alias{grep_match}
\title{Do GREP and MATCH in One Function}
\usage{
grep_match(sequ, dta, sort = FALSE)
}
\arguments{
\item{sequ}{vector of greps}

\item{dta}{list or vector to grep and match on}
}
\description{
Do GREP and MATCH in One Function.
}
\examples{
grep_match(c("^a", "^d", "^b"), c("house", "bird", "dog", "apple"))
}

