% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nunique.R
\name{nunique}
\alias{nunique}
\title{Find Length of Unique Terms in Vector}
\usage{
nunique(x)
}
\arguments{
\item{x}{vector}
}
\description{
Find Length of Unique Terms in Vector.
}
\examples{
nunique(c("A", "A", "B", "C")) # should = 3
}

