% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opener.R
\name{opener}
\alias{opener}
\title{Open (Directories & Files)}
\usage{
opener(x = ".")
}
\arguments{
\item{x}{A vector (typically of length one) of paths to directories of files.}
}
\description{
Use the operating system defaults to open directories and files.
}
\note{
This function is operating system and setting dependant.  Results may
not be consistent across operating systems.  Depending upon the default
programs for file types the results may vary as well.  Some files may not be
able to be opened.
}
\examples{
\dontrun{
opener()
}
}
\author{
Dason Kurkiewicz and Tyler Rinker <tyler.rinker@gmail.com>.
}
\references{
\url{http://stackoverflow.com/q/12135732/1000343}
}

