% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_lookup.R
\name{search_subfolders}
\alias{search_subfolders}
\title{Search Sub Folders}
\usage{
search_subfolders(subfolder_paths, colname, fun = NULL, pattern = ".csv$",
  ...)
}
\arguments{
\item{subfolder_paths}{A vector of directory paths to search in (from
\code{\link[base]{list.dirs}} or \code{get_subfolder_path}).}

\item{colname}{A column name to search for in the designated file type
(\code{pattern}).}

\item{fun}{A valiData validating function (optional).}

\item{pattern}{The file type to search for.}

\item{\ldots}{Additional arguments given to the \code{fun}.}
}
\value{
Gives data.frame and info on matching criteria.
}
\description{
Search subfolders for a file type with a particular column name and optionally validate.
}
\examples{
\dontrun{
if (!require("pacman")) install.packages("pacman")
pacman::p_load(magrittr, dplyr, cl)

## get csv files that have "AcademicStanding" header
potential_student_files <- cl::l_drive_go("/FTP") \%>\%
    list.dirs() \%>\%
    assign("dirs", ., .GlobalEnv) \%>\%
    ## get_subfolder_path("MyFolder") \%>\%
    cl::search_subfolders("AcademicStanding")

## filter results
potential_student_files \%>\%
    filter(perc_non_miss > 0) \%>\%
    print.data.frame()
}
}

